Brings list-based and SoF configs to the same Keycloak instance.
FIXES: b/214325255

TESTED:
Ran the docker compose file locally and tested the created Keycloak
instance against the proxy. Also tested the sample Growth Chart app
which could fetch the access token.

NOTE: there is currently a weird signature validation issue with
the SoF app which I will figure out and fix before submit and remove
this note! But this should be ready for review.

Change-Id: Iafdb252d1b29f7cec4ff2b8d8cbf70785712eee9